rename use-new-tag-mutation
