+++ client/src/components/NewTag/useNewTag.ts
@@ -4,7 +4,7 @@ import { useTagSelection } from "@lib/state/selected-tags-state";
 import useAuthentication from "@lib/use-authentication";
 import type { NewTag } from "@type/server/tag.types";
 import { useEffect, useState } from "react";
-import { useNewTagMutation } from "./use-new-tag-mutation";
+import { useNewTagMutation } from "./useNewTagMutation";
 
 export default function useNewTag() {
 	const { currentUser } = useAuthentication();
